Refactoring manager.go into separate concerns
